+++ CHANGELOG.md
@@ -7,6 +7,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Short windows now show the current synced lyrics line directly above the player bar.
+
 ## [0.29.0] - 2026-09-03
 
 ### Added